telegram-bot/requirements.txt
python-telegram-bot==21.10
requests==2.32.3
